+++ tsconfig.json
@@ -1,16 +1,14 @@
 {
-  "compilerOptions": {
-    "target": "ES2021",
-    "outDir": "dist",
-    "module": "commonjs",
-    "declaration": true,
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "strict": true,
-    "skipLibCheck": true,
-  },
-  "include": [
-    "src/**/*"
-  ],
-  "exclude": ["src/templates/dapp"]
+    "compilerOptions": {
+        "target": "ES2021",
+        "outDir": "dist",
+        "module": "commonjs",
+        "declaration": true,
+        "esModuleInterop": true,
+        "forceConsistentCasingInFileNames": true,
+        "strict": true,
+        "skipLibCheck": true
+    },
+    "include": ["src/**/*", "src/templates/dapp/src/utils/wrappersData.ts"],
+    "exclude": ["src/templates/dapp/**/*"]
 }
